make reset safer

+++ src/server_commands.py
@@ -5,7 +5,7 @@ class ServerSetup(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
-    @commands.command(aliases=['s'], description='create server instance')
+    @commands.command(aliases=['s'], description='create guild instance')
     async def setup(self, ctx):  # create new item in table
         if not db.guild_exists(str(ctx.guild.id)):
             db.create_guild(str(ctx.guild.id), str(ctx.channel.id))
@@ -16,12 +16,25 @@ class ServerSetup(commands.Cog):
     @commands.command(aliases=['r'], description='reset instance')
     async def reset(self, ctx):  # delete item from table
         if db.guild_exists(str(ctx.guild.id)):
-            db.destroy_guild(str(ctx.guild.id))
-            await ctx.message.add_reaction(u"\U0001F44D")
-        else:
-            await ctx.send('guild has not been setup')
+            await ctx.send('reset will clear all users, are you sure? y/n')
 
-    @commands.command(aliases=['rg'], description='change server region')
+            def check(m):
+                return m.author == ctx.author and m.channel == ctx.channel
+
+            try:
+                response = await self.bot.wait_for('message', check=check, timeout=10.0)
+            except Exception:
+                await ctx.send('timeout: guild not reset')
+            else:
+                if response.content.upper() == 'Y':
+                    db.destroy_guild(str(ctx.guild.id))
+                    await response.add_reaction(u"\U0001F44D")
+                elif response.content.upper() == 'N':
+                    await ctx.send('guild not reset')
+        else:
+            await ctx.send('guild not setup')
+
+    @commands.command(aliases=['rg'], description='change guild region')
     async def region(self, ctx, arg=None):  # set new region of item in table
         if db.guild_exists(str(ctx.guild.id)):
             regionlist = ['BR', 'EUNE', 'EUW', 'JP', 'KR',
@@ -31,7 +44,7 @@ class ServerSetup(commands.Cog):
                 return
             current_region = arg.upper()
             if current_region in regionlist:
-                await ctx.send('changing server region will clear all users, are you sure? y/n')
+                await ctx.send('changing guild region will clear all users, are you sure? y/n')
 
                 def check(m):
                     return m.author == ctx.author and m.channel == ctx.channel
@@ -39,7 +52,7 @@ class ServerSetup(commands.Cog):
                 try:
                     response = await self.bot.wait_for('message', check=check, timeout=10.0)
                 except Exception:
-                    await ctx.send('timeout: server region not changed')
+                    await ctx.send('timeout: guild region not changed')
                 else:
                     if response.content.upper() == 'Y':
                         updates = {
@@ -49,7 +62,7 @@ class ServerSetup(commands.Cog):
                         db.update_guild(str(ctx.guild.id), updates)
                         await response.add_reaction(u"\U0001F44D")
                     elif response.content.upper() == 'N':
-                        await ctx.send('server region not changed')
+                        await ctx.send('guild region not changed')
             else:
                 await ctx.send('region not found')
         else:
